Fixes and enhancements
