+++ testfm/backup.py
@@ -25,3 +25,39 @@ class Backup(Base):
     """Manipulates Foreman-maintain's backup command"""
 
     command_base = 'backup'
+
+    @classmethod
+    def run_online_backup(cls, options=None):
+        """Build foreman-maintain backup online"""
+
+        cls.command_sub = 'online'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+        return result
+
+    @classmethod
+    def run_offline_backup(cls, options=None):
+        """Build foreman-maintain backup offline"""
+
+        cls.command_sub = 'offline'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+        return result
+
+    @classmethod
+    def run_snapshot_backup(cls, options=None):
+        """Build foreman-maintain backup snapshot"""
+
+        cls.command_sub = 'snapshot'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+        return result
