+++ testfm/upgrade.py
@@ -24,3 +24,39 @@ class Upgrade(Base):
     """Manipulates Foreman-maintain's health command"""
 
     command_base = 'upgrade'
+
+    @classmethod
+    def list_versions(cls, options=None):
+        """Build foreman-maintain upgrade list-versions"""
+        cls.command_sub = 'list-versions'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+
+        return result
+
+    @classmethod
+    def check(cls, options=None):
+        """Build foreman-maintain upgrade check"""
+        cls.command_sub = 'check'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+
+        return result
+
+    @classmethod
+    def run(cls, options=None):
+        """Build foreman-maintain upgrade run"""
+        cls.command_sub = 'run'
+
+        if options is None:
+            options = {}
+
+        result = cls._construct_command(options)
+
+        return result
